requirements.txt
openai>=1.27.0,<2.0.0
python-dotenv>=0.21.0
pandas>=1.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
click>=8.1.0
Flask
